Measure the runtime

+++ python_async_function/2-measure_runtime.py
@@ -10,4 +10,4 @@ async def measure_time(n: int, max_delay: int) -> float:
     start = asyncio.get_event_loop().time()
     await wait_n(n, max_delay)
     end = asyncio.get_event_loop().time()
-    return (end - start) / n
+    return float((end - start) / n)
